backend/requirements.txt
django-stubs==4.2.7
Django==5.0.7
djangorestframework==3.15.2
drf-spectacular==0.27.2
gunicorn==22.0.0
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
pydantic==2.5.2
whitenoise==6.6.0
